Time: 0 ms (100.00%), Space: 6 MB (48.74%) - LeetHub

+++ 191-number-of-1-bits/191-number-of-1-bits.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int hammingWeight(uint32_t n) {
-        int count = 0;
+        uint8_t count = 0;
         while(n) {
             count += n & 0x1 ? 1 : 0;
             n = n >> 1;
